platform/property/src/test/java/org/jage/property/monitors/PropertyMonitorRulePairTest.java
/**
 * Copyright (C) 2006 - 2012
 *   Pawel Kedzior
 *   Tomasz Kmiecik
 *   Kamil Pietak
 *   Krzysztof Sikora
 *   Adam Wos
 *   Lukasz Faber
 *   Daniel Krzywicki
 *   and other students of AGH University of Science and Technology.
 *
 * This file is part of AgE.
 *
 * AgE is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * AgE is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with AgE.  If not, see <http://www.gnu.org/licenses/>.
 */
package org.jage.property.monitors;

import static org.junit.Assert.assertSame;

import org.jage.property.testHelpers.PropertyMonitorStub;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.runners.MockitoJUnitRunner;

@RunWith(MockitoJUnitRunner.class)
public class PropertyMonitorRulePairTest {

	@Mock
	IPropertyMonitorRule rule;
	
	@Test
	public void testMonitorRulePair() {
		AbstractPropertyMonitor monitor = new PropertyMonitorStub();
		
		PropertyMonitorRulePair pair = new PropertyMonitorRulePair(monitor, rule);
		
		assertSame(monitor, pair.getPropertyMonitor());
		assertSame(rule, pair.getPropertyMonitorRule());
		
	}
}
